use manifests configmap to simulate bootstrap failure
Now that Hive has a way for users to supply their own manifests, there
is no need to use an annotation to simulate a bootstrap failure. Instead
of using an annotation, the bad manifest can be added directly using the
manifests configmap.

The --simulate-bootstrap-failure option for hiveutil create-cluster has
been retained. Its behavior has been changed from adding an annotation
to the clusterdeployment to adding a file in the manifests configmap.