dnsendpoint: do not reconcile until initial scrape of root domain
When the controllers start, all existing dnsendpoint resources are queued
for reconciliation. This happens concurrently with the initial scrape of
the name servers. If a dnsendpoint is reconciled prior to the initial scrape,
then the controller will determine that the NS records for the dnsendpoint
do not exist and attempt to create them. This causes unnecessary and bursty
traffic with the cloud provider. These changes force the controller to wait
until the initial scrape is complete before reconciling any dnsendpoint resources.

Note that reconciling a deleted dnsendpoint is allowed. We want to allow the
delete to go through even if there has not been a scrape so that deletes can
still happen in the event of a problem with the scraper.