Add basic support for bare metal cluster provisioning.
Adds a new platform, currently just an empty struct, for BareMetal.

If we're provisioning bare metal we need to lookup a ref for a different
installer image. Using this image the openshift-install binary will be
in the exact same location with the exact same name as we normally use,
so we just need to swap out the image reference and everything else can
remain the same.

There is a pending need for installing mkisofs which is hopefully going
to be handled by the requirement being removed upstream soon.