Add SyncSet metrics.
Adds four new metrics, and SyncSetInstance.Status.Applied for a simpler
API to determine if everything in the SyncSet is applying properly.
Currently you must interpret at least 4 conditions, some of which are
errors and one of which is success. The new field is displayed as a
column in kubectl get output.

hive_selectorsyncset_clusters_total is a gauge labeled by name showing
the number of clusters that matched the label selector.
hive_selectorsyncset_clusters_unapplied_total is similar but for the
number that are not currently applying. Both are calculated by
examining SyncSetInstances.

hive_syncsets_total is a gauge that shows the total number of SyncSets,
and hive_syncsets_unapplied_total the number that are not currently
applying. This is also calculated by examining SyncSetInstances.