installmanager: fix interactions with installer logs
https://jira.coreos.com/browse/CO-656
The installer output was being redirected to a newly created
file each time that the installer was executed via the
runOpenShiftInstallCommand method. This changes the installmanager
so that the destination file is appended instead of truncated,
thus preserving the outputs for every installer execution.

https://jira.coreos.com/browse/CO-589
The full installer logs were being piped to the output of the
container every time that the installer was executed via the
runOpenShiftInstallCommand method. This changes the installmanager
so that the piping is set up once at the start of the process instead
of with each installer execution.