allow managing dns for multiple cloud providers simultaneously
update hiveconfig to allow defining multiple entries for each managed domain where each entry can contain a list of domains and any cloud-specific data (ie credentials) for managing those domains.

update the hive operator to send a JSON representation of the hiveconfig.spec.managedDomains fields as the configmap to be mounted for the dnsendpoint controller.

update hiveutil to process updates to hiveconfig.spec.managedDomains.

modify the dnsendpoint controller to read the new format of the file containing the managed domains configuration (now just a JSON dump of hiveConfig.Spec.ManagedDomains).

modify ReconcileDNSEndpoint struct to maintain a list of domains being managed.

add test cases to ensure proper processing of the updated configuration file format.

update existing test cases.

add DomainNotManagedCondition. this can happen in the event that a DNSZone object that was created when the hiveConfig has support for domain x.y.z, but an updated hiveConfig no longer has a config for x.y.z. Set a condition on the DNSZone to note this. also, allow dnsendpoint controller to always run so that it can set the DomainNotManaged condition when HiveConfig has no ManagedDomains configuration

handle updating an existing NS entry in GCP. when restarting hive, the dnsendpoint controller will re-reconcile all DNSZone objects. this leads down the path of creating/updating the DNS entries out in GCP as needed. the GCP nameserver code (for the dnsendpoint controller) would attempt to recreate the NS records, and that would error and bubble up back to the dnsendpoint controller, causing us to never complete reconciliation of the DNSZone. add a new UpdateResourceRecordSet call to the GCP client, and use it to allow the update path to succeed.

add live test case to cover the new update path.