types: remove DefaultMachinePlatform
The use of this field was lost in either the change to v1 or the change
to the MachinePool CRD. It was never working entirely correctly. It does
not seem too onerous to require the user to fill out each MachinePool fully
rather than relying on ClusterDeployment-wide defaults. So, these changes
remove the DefaultMachinePlatform fields from ClusterDeployment.

https://jira.coreos.com/browse/CO-580