Fix baremetal provisioning when run on OpenShift.
This was fine in Kube but in OpenShift both ssh to libvirt, and .cache
image copying broke due to the security constraints in OpenShift.

This change adds a home dir, changes it's permissions so it will work
with a random UID, and adds that random runtime UID to /etc/passwd so
SSH will work.

Also moved mounting of the ssh private key such that it will always
happen even if log gathering is disabled.