api url override
Add a feature for overriding the API URL that Hive uses to communicate
with the remote cluster. This is set via the
.spec.controlPlaneConfig.apiURLOverride field in the ClusterDeployment.

When the field is set, the unreachable controller will attempt to connect
to the override URL when checking for whether the remote cluster is
reachable. If the override URL is reachable, then all other controllers
that connect to the remote cluster will use the override URL. If the
override URL is not reachable, then the unreachable controller will
attempt to connect to the initial API URL set when the remote cluster
was installed. If the initial URL is reachable, then all other controllers
will use the initial URL when connecting to the remote cluster. If both
the override URL and the initial URL are unreachable, then the unreachable
controller will mark the remote cluster as unreachable. No other controllers
will attempt to connect to the remote cluster in this case.

The "Unreachable" condition is true when both the initial URL and the
override URL are unreachable. The message will include the errors from
both connection attempts.

The "ActiveAPIURLOverride" condition is true when the override URL
is set and reachable. If the condition is false, the message will include
the error from the connection attempt using the override URL.

https://jira.coreos.com/browse/CO-669