apiserver: block v1alpha1 hiveconfig updates when multiple managed domains
The v1alpha1 HiveConfig only supports a single cloud for managed DNS.
If the v1 HiveConfig has multiple entries in its .spec.managedDomains
field, then the v1alpha1 update will lose all but one of the entries,
even if the update was updating a different field. To prevent this,
the v1alpha1 update will fail if the HiveConfig has multiple entries
in its .spec.managedDomains field.