Implement support for multiple MachinePools on GCP.
Due to some extreme naming conventions on GCP, we have effectively one
character to play with in terms of the resulting MachineSets and Machine
in-cluster.

To accomodate this we implement a single character leasing system to
automatically assign a character to each new machine pool, and ensure
that only that pool can have that character.

'w' is reserved for the default "worker" pool from the installer to
prevent cycling all existing nodes. 'm' is similarly reserved as it was
used for the masters.

Other pools will have their lease character chosen at random.