Extract URLs from kubeconfig for adopted clusters without a provision.
Adopted clusters do not have a ClusterProvision, but this code was the
only place we set Status.APIURL and Status.WebConsoleURL.

Will now ensure these are set even if the ClusterDeployment is marked
installed.

The ClusterProvision processing still has similar code just to avoid an
extra status update on intial processing of a completed install.