handle changing managed domains in hiveConfig
The hive operator would not relaunch the hive-controllers deployment when the only change to the HiveConfig was adding/removing an existing managed domain entry (leaving at least one managed domain).

Update the logic to create a new configmap with a new name (so that the controllers cannot see the old configmap contents on controller restart), and detect when changes are needed so that the hive-controllers deployment is re-deployed as necessary.

A notable change is that a HiveConfig with no managed domain entries will now have a configmap with an empty list of managed domains used in the hive-controllers deployment (the dns controllers already handle the case of an empty list of managed domains). Previously the hive-operator would simply not create the configmap at all.

Handle migrating away from the old static configMap name of "managed-domains" so that the legacy configMap isn't orpahned.

Add status fields for HiveConfig: Add observedGeneration and configApplied fields to HiveConfig.Status.; Update hive-operator to record the last reconciled Generation of the HiveConfig and whether or not it was successful.

Modify the 'hiveutil adm managed-dns enable' command to key off of these new status fields for synchronizing around whether the new manage-dns settings have rolled out.