Formalize addition of ssh known hosts in install manager.
Originally done as part of a PoC for bare metal provisioning with a
dedicated libvirt hypervisor host, we now wish to make this a more long
term hive feature.

As a first step in this, this change removes the ssh known host
annotation and makes it a formal part of the Hive API. The hosts are now
represented in in spec.Provisioning.SSHKnownHosts. This is generic and
not tied to bare metal provisioning as we're not sure the extent we may
need ssh in installs in the future. We already use ssh to gather logs on
failures and while this does not currently require known hosts entries,
it seems best to keep the door open.