Separate cluster ssh key from libvirt connection ssh key.
For PoC purposes we assumed to use the same ssh key for dropping on the
cluster hosts as we do for connecting to libvirt for bare metal
provisioning. There is no good reason to do this so for a medium term
solution this change adds a new SSH key that can be specified for
libvirt provisioning.