apiserver: fix rbac and port used
Change the port that the apiserver is listening on to 10443 to match the
target port set for the service.

Add a ClusterRoleBinding for the apiserver service account to system:auth-delegator.

Remove the code in hive-operator that was trying to set up for RBAC for the apiserver
service account. This will always fail because the operator does not have the necessary
permissions to pass to the apiserver. The RBAC needs to be set up for the apiserver in
the same way that it is set up for the hiveadmission.