Remove now obsolete toleration edit in remoteingress controller.
This was done to force the rollout of ingress config, which should no
longer be needed as of 4.2.0.

The code in question is now causing a problem for hive users who need to
adjust default router tolerations to land them on specific nodes, also
being done by syncsets, and now both syncsets are fighting for the
tolerations causing the pods to restart often, impacting running
services.

With this change, remoteingress will no longer work seamlessly on 4.1
OpenShift clusters, manual restarts may be required. We do not believe
this will impact many users as 4.2 and 4.3 are both available now.