support aws china for managed dns
Use a custom endpoint for route53 when the region is a "cn-" region.

A region field has been added to .spec.managedDomains[].aws in HiveConfig.
When the managed domain is handled by AWS China, the region field should
be set to "cn-northwest-1" so that the route53 requests are sent to AWS
China instead of regular AWS.

https://issues.redhat.com/browse/CO-769