Add ClusterPool API and controller.
Allows defining a pool of clusters of a certain size to be kept ready
for use. Uses the builder we use for hiveutil create-cluster to generate
all artifacts, including an install-config. As such options for
customization are a bit restricted, but should be helpful for anyone
looking to keep clusters up and running.

Future work will allow clusters to be claimed by users, and immediately
hibernating clusters as soon as they're ready and waking them up when
claimed.