Fix bug in GCP machine pool name leases.
Late change to filter existing leases on those for a pool hide the rest
of the pool leases from the code that calculates what characters are
available. This resulted in collisions in some situations.

Modified to list all leases for the cluster, check if any are for our
specific pool, and if not iterate everything returned to find the in-use
lease chars.