update cleanupLogOutput to handle malformed data from libvirt and ssh
When the baremetal installer fails to connect via ssh to a remote
libvirt host, the error message included embedded and *escaped*
carriage returns and newlines. This change unescapes them so that any
redaction of the error message contents is minimized. The change is in
cleanupLogOutput instead of the function that reads the log because
there are no tests for the reading function and there are for the
cleaning function.

Also update TestCleanupRegex to show the unexpected output when the
regex fails to strip text.

Addresses #878

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>