Fix crash when using more than 35 GCP MachinePools.
Code did not properly handle the case when no lease characters are
available.

Added a short docs blurb, and a condition to be set when this situation
arises.

Lease deletions are now watched and will requeue all pools for the
cluster in case any are stuck without a lease character.