apiserver: omit empty serviceNetwork and machineNetwork from installconfig
The aggregated API server was assuming that the serviceCIDR and machineCIDR were
required fields in the v1alpha1 ClusterDeployment. However, they can be omitted
to use defaults. When they were omitted in the v1alpha1 ClusterDeployment,
the aggregated API server was creating a bad element in the serviceNetwork
and machineNetwork fields in the installconfig.

These changes
 (1) keep the bad element from being created and
 (2) treat a bad element (with the text "<nil>") as non-existent to correct the
     problem for existing clusters.