Improved clearing of provision/deprovision underway metrics.
Situations were encountered where the edge-trigger event for clearing
the metrics was being missed, resulting in alerts that would never clear
until the pod was restarted.

Delete was discovered and allows us to clear a metric, regardless if it
exists or not. This is far preferable to setting to 0 in very specific
scenarios and then stopping reporting.