admission: use kube decoder instead of go json unmarshal
The basic GO JSON unmarshal is case-insensitive for field names. Kube
is case-sensitive for field names. Since the admission webhooks were
using JSON unmarshal, resources that used the incorrect case for fields
were being incorrectly accepted. For example, if a user attempted to
create a bare-metal ClusterDeployment but used .spec.platform.baremetal
instead of .spec.platform.bareMetal, then the ClusterDeployment would
be accepted by the validating webhook even though it should be rejected
for not having a platform specified.

https://issues.redhat.com/browse/CO-846