Detect serving cert changes in operator instead of hiveadmission.
Trying an alternate approach for detecting changes to the serving cert
by monitoring the secret in the hive-operator, and on any change we
reconcile. Reconcile will then hash the secret data and store in an
annotation on the pod template in the hiveadmission deployment. Any
change to this value will result in a new rollout.