add new operator bundle generate script for operatorhub deploys
Add new generate-operator-bundle-operatorhub.py script. This script
is adapted from the existing hack/generate-operator-bundle.py, and
is capable of generating a bundle for both operatorhub and OSD. For
now we're only using it for operatorhub and keeping the OSD pipeline
unchanged. At a later date we'll adapt the OSD pipeline and rename
this script.