Update for developing on kind 0.8.1.
Switch to a much simpler script for creating a kind cluster to develop
hive against, with an integrated registry. The registry is now linked to
the container with docker networking and the new script takes place of
both of the old ones, if there is no registry container one will be
started for you.