move cluster to another hive instance
When a ClusterRelocate has a label selector that matches with
a ClusterDeployment, the clusterrelocate controller will relocate
the matching ClusterDeployment.

1. Set the hive.openshift.io/relocating annotation on the ClusterDeployment.
2. Copy secrets, configmaps, machinesets, syncsets, and syncidentityproviders.
3. Copy the DNSZone for the ClusterDeployment.
4. Copy the ClusterDeployment.
5. Replace the hive.openshift.io/relocating annotation with the
   hive.openshift.io/relocated annotation.

When a ClusterDeployment has either the relocating or relocated annotation,
no controllers will do any mutation of the remote cluster (such as syncing
syncsets). When the ClusterDeployment has the relocated annotation, then
controllers will not run finalizer code when the ClusterDeployment is deleted.

https://issues.redhat.com/browse/CO-653