move cluster to another hive instance
When a ClusterRelocate has a label selector that matches with
a ClusterDeployment, the clusterrelocate controller will relocate
the matching ClusterDeployment.

1. Set the hive.openshift.io/relocate annotation to outgoing on the
   ClusterDeployment and the DNSZone.
2. Copy secrets, configmaps, machinesets, syncsets, and syncidentityproviders.
3. Copy the DNSZone for the ClusterDeployment, with the relocate annotation
   set to incoming.
4. Copy the ClusterDeployment, with the relocate annotation set to incoming.
5. Set the relocate annotation on the DNSZone and ClusterDeployemnt to complete.

When a ClusterDeployment with a relocate annotation set to incoming is
reconciled by the clusterRelocate controller, the relocate annotation is
removed from the DNSZone and the ClusterDeployment. This indicates that the
relocate has completed on the destination side.

When a ClusterDeployment has the relocated annotation, no controllers will do
any mutation of the remote cluster (such as syncing syncsets). When the
ClusterDeployment has the relocated annotation, then controllers will not run
finalizer code when the ClusterDeployment is deleted.

Add a hive_cluster_relocations counter vec that tracks the total number
of successful cluster relocations. The metric has a single
cluster_relocate label that is the name of the ClusterRelocate
directing the relocation.

Add a hive_aborted_cluster_relocations counter vec that tracks the
total number of aborted cluster relocations. The metrics has two
labels: cluster_relocate and reason. The cluster_relocate label
is the name of the ClusterRelocate directing the aborted relocation.
The reason label is the reason why the relocate was aborted. Possible
values for the reason label are "no_match", "multiple_matches", and
"new_match".

The number of failing cluster relocates can be seen by looking at
the hive_cluster_deployments_conditions metric and filtering on
a value of RelocationFailed for the condition label.

https://issues.redhat.com/browse/CO-653