Add MachinePool support for bring your own VPC
Currently, Hive can install the cluster in an already existent VPC. But when a
new MachinePool is created, Hive creates the MachineSets in the cluster without
specifying the subnet IDs, thus causing the Machine objects to get into failed
state. This commit adds an option to specify the subnet IDs to be used in the
MachinePool.