protect against accidental deletes of installed clusters
The admission webhook for CluterDeployments has been modified to inspect
DELETE requests as well. The webhook will reject any DELETE request for
a ClusterDeployment that has the "hive.openshift.io/protected-delete" annotation
set to a true value. In order to delete such a ClusterDeployment, a user
is required to delete the annotation prior to making the DELETE request.

The "deleteProtection" field has been added to the HiveConfig. When the
field is set to "enabled", the clusterdeployment controller will add the
"hive.openshift.io/protected-delete" annotation to the ClusterDeployment
when transitioning the ClusterDeployment to installed.

https://issues.redhat.com/browse/CO-277