controller: add metrics for cluster relocations
Add a hive_cluster_relocations gauge vec that tracks the total number
of successful cluster relocations. The metric has a single
cluster_relocate label that is the name of the ClusterRelocate
directing the relocation.

Add a hive_aborted_cluster_relocations gauge vec that tracks the
total number of aborted cluster relocations. The metrics has two
labels: cluster_relocate and reason. The cluster_relocate label
is the name of the ClusterRelocate directing the aborted relocation.
The reason label is the reason why the relocate was aborted. Possible
values for the reason label are "no_match", "multiple_matches", and
"new_match".

The number of failing cluster relocates can be seen by looking at
the hive_cluster_deployments_conditions metric and filtering on
a value of RelocationFailed for the condition label.