Disable DNSZone relocation temporarily.
We are seeing stuck terminating namespaces in some of our environments,
it appears it is possible for the ClusterDeployment to be gone when
reconciling a deleted DNSZone, despite the owner reference with
blockOwnerDeletion.

Add metrics to track the force deletion of dnszones which was quietly
being handled despite possibly leaking resources.