Fix unable to install CRD and CR in same syncset
Currently, we first gather info about all the resources before creating them.
This way we cannot create a CRD and CR in the same syncet, because to fetch the
info about the CR we need to first create the CRD. This commit fixes it by
fetching info and creating one resource at a time.