Add syncset time to apply metrics
This commit adds the metric that shows the time to apply syncsets and
selectorsyncsets. The time is logged in seconds and is calculated when
the syncsetinstance is applied successfully for the first time.If the
syncsetinstance is spawned from a selectorsyncset, it will report
the metric along with the selectorsyncset name as the label.