add support for full gcp machine pool names
Starting with OpenShift 4.4.8, the restriction that machine pool names
can only be a single character long has been lifted. As a consequence,
the worker machinesets created by the installer now use the full name
"worker", where Hive assumes that they will use the name "w".

These changes will use the full pool name when the cluster has a version
of 4.5 or greater and continue to use name leases for clusters with
older versions. However, for clusters that started as older versions
and have been upgraded to 4.5 or greater, name leases will be used if
there are existing leases or if Hive determines that the worker machinesets
created by the installer--which used the "w" pool name--still exist.

https://issues.redhat.com/browse/CO-1030