fix update to set cluster version labels
In https://github.com/openshift/hive/pull/1041, the clusterversion controller
was changed to set labels on the ClusterDeployment for the cluster version.
However, the controller was not changed to do a regular update of the
ClusterDeployment. The only update remained a Status update. The changes
here separate the code to set the labels into a new function in which a
regular update of the ClusterDeployment is executed if there are any changes
to the labels.

https://issues.redhat.com/browse/CO-626