reap namespaces created for ClusterPool after clusters deleted
When the ClusterPool creates a new cluster, it creates a namespace to
house the ClusterDeployment. When the ClusterDeployment is deleted,
the namespace then needs to be deleted to avoid leaking namespaces.
To this end, these changes add a clusterpoolnamespace controller which
will sync namespaces and delete them when appropriate. All namespaces
created for ClusterPools will have a hive.openshift.io/origin-cluster-pool-name
label. The new controller will delete a namespace with that label if
the namespace does not contain any ClusterDeployments. The controller
will wait 5 minutes after the creation of the namespace before
deleting the namespace to give time for the ClusterDeployment to be
created.

https://issues.redhat.com/browse/CO-1014