stop adding MachinePool twice
With https://github.com/openshift/hive/pull/1024, code was added
to the create-cluster command to skip the MachinePool for certain
platforms. However, the code that added the MachinePool
indiscriminately was not removed. The result was that 1 MachinePool
was added for some platforms and 2 duplicates MachinePools were
added for other platform.