respect the region specified in the clusterpool
Plumb through a Region field in the AWS, Azure, and GCP CloudBuilders.
Set the region in the CloudBuilder from the region specified in the
ClusterPool so that the clusters created for the pool are created in
the desired region.

Additionally, add the --region flag to the `hiveutil create-cluster`
command that will dictate in which region the cluster is created. If
the --region flag is used for a platform other than AWS, Azure, or
GCP, then the command will fail.

https://issues.redhat.com/browse/CO-801