add expectations to clusterpool controller
Add functionality to the clusterpool controller to wait for
ClusterDeployments to be seen in the store after creation before
doing any further reconciliation of the ClusterPool. This is
needed to prevent too many ClusterDeployments from being created
as a result of the clusterpool controller seeing an updated
ClusterPool before the newly created ClusterDeployment has been
seen.