add clusterclaim type
A ClusterClaim resource is created by a user when they want to claim a
cluster from a ClusterPool. The ClusterClaim must be in the same namespace
as the ClusterPool. When a cluster has been assigned to the ClusterClaim,
the namespace field in the status will be filled out with the namespace
where the ClusterDeployment for the cluster resides, and the NotReady
condition will be set to False.

https://issues.redhat.com/browse/CO-1012