assign clusters to clusterclaims
When reconciling a ClusterPool, the clusterpool controller will check for
ClusterClaims that reference the ClusterPool and that do not have a cluster
assigned. If there are ready clusters, the controller will set the
`.status.namespace` field to the namespace of one of the ready clusters.
If there are not enough ready clusters to fulfill all claims, the controller
will prioritize the oldest claims. If a claim is not assigned a cluster,
the NotReady condition will be set to True on the claim with a reason of
NoClusters.

The clusterclaim controller has been added to handle reconciling ClusterClaims
that have been assigned a cluster. For a claim with an assignment, the controller
will check the ClusterDeployment to ensure that the cluster has not been assigned
to a different claim. If it has, the the `.status.namespace` field of the
ClusterClaim will be cleared. If the ClusterDeployment has not been associated with
any claim, then the controller will update the ClusterDeployment to associated it
with the ClusterClaim, completing the cluster assignment. When the ClusterClaim is
deleted, the controller will delete the ClusterDeployment assigned to the claim.

https://issues.redhat.com/browse/CO-1012