create RBAC for accessing claimed cluster
When a ClusterClaim is assigned a cluster, the clusterclaim
controller will create a Role and RoleBinding in the namespace
of the ClusterDeployment. The Role will grant full permissions for
Hive resources and GET permissions for the admin kubeconfig and
password secrets. The RoleBinding will assign the role to the
subjects specified in the ClusterClaim.

https://issues.redhat.com/browse/CO-1070