code review feedback
* Moved the `.status.namespace` field of `ClusterClaim` to
`.spec.namespace`. This preserves the cluster assignment if the
`ClusterClaim` loses its status, say due to a restore from backup.
* Changed terminology around cluster assignment. When the clusterpool
controller sets the `.spec.namespace` field of a `ClusterClaim`,
that is "assigning" the cluster to the `ClusterClaim`. When the
clusterclaim controller sets the `.spec.clusterPoolRef.claimName`
field of a `ClusterDeployment`, that is "claiming" the cluster.