validation: base domain that is a managed domain
When one managed domain is a child of another managed domain,
the the user can create a ClusterDeployment with a base domain
that is the same as the child managed domain. This will cause
problems when it comes time to change the recordsets in the
base domain. These changes will have the validating webhook
reject the ClusterDeployment.

For example, if there are managed domains of example.com and
child.example.com, then the user should not be allowed to
create a ClusterDeployment with a base domain of child.example.com.