apis: add clustersync type
Create a new hiveinternal.openshift.io/v1alpha1 API group.

Add the ClusterSync type to the new API group. This type will be
used to track the SyncSets and SelectorSyncSets that are applied
to a cluster.

Add the ClusterSyncLease type to the new API group. This type
will be used to track when the last time the SyncSets and
SelectorSyncSets were applied to a cluster. This is tracked
separately from the ClusterSync type so that, during the happy
path scenario, the periodic update of syncsets does not require
an update of a ClusterSync resource, which can be very large.